requirements.txt
flask
vk
psycopg2
pyyaml
pandas
sklearn